refactor: delete RocketCardView and replace all logic to Cell
