+++ SpaceX.xcodeproj/xcuserdata/egbad.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -20,5 +20,113 @@
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675"
+            shouldBeEnabled = "No"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "13"
+            endingLineNumber = "13"
+            landmarkName = "RocketsCollectionViewCell"
+            landmarkType = "3">
+            <Locations>
+               <Location
+                  uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675 - a1282169ef22939b"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "SpaceX.RocketsCollectionViewCell.rocket.getter : Swift.Optional&lt;SpaceX.Rocket&gt;"
+                  moduleName = "SpaceX"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/egbad/Developer/SpaceX/SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "13"
+                  endingLineNumber = "13"
+                  offsetFromSymbolStart = "2964">
+               </Location>
+               <Location
+                  uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675 - 127dbba026e12f97"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "SpaceX.RocketsCollectionViewCell.rocket.setter : Swift.Optional&lt;SpaceX.Rocket&gt;"
+                  moduleName = "SpaceX"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/egbad/Developer/SpaceX/SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "13"
+                  endingLineNumber = "13"
+                  offsetFromSymbolStart = "3004">
+               </Location>
+               <Location
+                  uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675 - 271f54f4c676f593"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "SpaceX.RocketsCollectionViewCell.rocket.modify : Swift.Optional&lt;SpaceX.Rocket&gt;"
+                  moduleName = "SpaceX"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/egbad/Developer/SpaceX/SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "0"
+                  endingLineNumber = "0"
+                  offsetFromSymbolStart = "27">
+               </Location>
+               <Location
+                  uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675 - ff9f5bcc240f3ca6"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "SpaceX.RocketsCollectionViewCell.rocket.modify : Swift.Optional&lt;SpaceX.Rocket&gt; with unmangled suffix &quot;.resume.0&quot;"
+                  moduleName = "SpaceX"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/egbad/Developer/SpaceX/SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "13"
+                  endingLineNumber = "13"
+                  offsetFromSymbolStart = "15">
+               </Location>
+               <Location
+                  uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675 - f03978b0e0c85d5"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "SpaceX.RocketsCollectionViewCell.init(frame: __C.CGRect) -&gt; SpaceX.RocketsCollectionViewCell"
+                  moduleName = "SpaceX"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/egbad/Developer/SpaceX/SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "13"
+                  endingLineNumber = "13"
+                  offsetFromSymbolStart = "200">
+               </Location>
+               <Location
+                  uuid = "00A7BC10-5F48-4F49-A9D6-8BEC81053675 - 3f82a56fe610cabd"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "SpaceX.RocketsCollectionViewCell.init(coder: __C.NSCoder) -&gt; Swift.Optional&lt;SpaceX.RocketsCollectionViewCell&gt;"
+                  moduleName = "SpaceX"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/egbad/Developer/SpaceX/SpaceX/Views/Cell/RocketsCollectionViewCell.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "13"
+                  endingLineNumber = "13"
+                  offsetFromSymbolStart = "104">
+               </Location>
+            </Locations>
+         </BreakpointContent>
+      </BreakpointProxy>
    </Breakpoints>
 </Bucket>
