+++ SpaceX/Models/RocketCellData.swift
@@ -1,0 +1,13 @@
+//
+//  RocketCellData.swift
+//  SpaceX
+//
+//  Created by Егор Бадмаев on 10.03.2022.
+//
+
+struct RocketCellData {
+    let name: String
+    let imageURL: String
+    let flightDate: String
+    let 
+}
